kiosk_app/requirements.txt
Flask==3.0.3
Flask-Babel==4.0.0
python-dotenv==1.0.1
sentry-sdk==2.17.0
supabase==2.9.1
tenacity==9.0.0
